requirements.txt
flask
flask-socketio
mistralai
python-dotenv
SQLAlchemy
mysql-connector-python
